+++ src/main/java/com/crowdplatform/service/BatchService.java
@@ -31,14 +31,14 @@ public interface BatchService {
      * @param projectId The project id
      * @return List with the ids for the batches with status RUNNING
      */
-    public List<Integer> listRunningBatchIds(Integer projectId);
+    public List<Integer> listRunningBatchIds(Long projectId);
 	
     /**
      * Retrieve all ids for batches in the system with a COMPLETED status for a given project.
      * @param projectId The project id
      * @return List with the ids for the batches with status COMPLETED
      */
-    public List<Integer> listCompletedBatchIds(Integer projectId);
+    public List<Integer> listCompletedBatchIds(Long projectId);
     
     /**
      * Set a batch's status to RUNNING.
